+++ src/js/app.js
@@ -9,6 +9,7 @@ angular.module('brewbench-monitor', [
   ,'duScroll'
   ,'ui.knob'
   ,'rzModule'
+  ,'ngRaven'
 ])
 .config(function($stateProvider, $urlRouterProvider, $httpProvider, $locationProvider, $compileProvider) {